pyproject.toml
[tool.poetry]
name = "wizbot"
version = "0.1.0"
description = ""
authors = ["Valentin B. <valentin.be@protonmail.com>"]
license = "ISC"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiosqlite = "^0.19.0"
"discord.py" = "^2.2.3"
jishaku = "^2.5.1"
loguru = "^0.7.0"
uvloop = { version = "^0.17.0", markers = "sys_platform != 'win32'" }
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pre-commit = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bot = "wizbot.__main__:main"
